+++ hrpGeplib/hrpGepModel/Link.h
@@ -175,15 +175,13 @@ namespace hrpGep {
         double gearRatio;
         double gearEfficiency;
         double rotorResistor;
-        /*
+
 		//Motor Friction Parameters
 		double Kv_p;
 		double Kv_n;
-		double Ka_p;
-		double Ka_n;
 		double Kf_p;
 		double Kf_n;
-        */
+
         bool isHighGainMode;
         bool isCrawler;
